feat: implements method 'findTotalNumberOfRegisteredMovies'

+++ trabalho02/src/main/java/trabalho02/dao/FilmeDao.java
@@ -19,7 +19,7 @@ public interface FilmeDao {
 
 	List<String> findTitlesOfAllMoviesContaining(String substring);
 
-	int findTotalNumberOfRegisteredMovies();
+	long findTotalNumberOfRegisteredMovies();
 
 	Filme save(Filme filme);
